refactor(api): extract cache + backup-routes, handlers.ts becomes pure shim (#35)
